fix: octokit, doesRepositoryExistsInGitHub() を別ファイルに切り出し
