+++ src/lib/does-repository-exists-in-github.ts
@@ -1,0 +1,11 @@
+import { octokit } from "./octokit";
+
+export const doesRepositoryExistInGitHub = async (repositoryUrl: string) => {
+  try {
+    const { data } = await octokit.request("GET " + repositoryUrl);
+    return data ? true : false;
+  } catch (error) {
+    console.error(error);
+    return false;
+  }
+};
