+++ src/lib/octokit.ts
@@ -1,0 +1,3 @@
+import { Octokit } from "octokit";
+
+export const octokit = new Octokit({ auth: process.env.GITHUB_TOKEN });
